Do proper validation for various areas
