+++ src/SeoToolkit.Umbraco.SiteAudit.Core/Models/PostModels/CreateAuditPostModel.cs
@@ -1,10 +1,13 @@
 ﻿using System;
+using System.ComponentModel.DataAnnotations;
 
 namespace SeoToolkit.Umbraco.SiteAudit.Core.Models.PostModels
 {
     public class CreateAuditPostModel
     {
+        [Required]
         public string Name { get; set; }
+        [Required]
         public Guid SelectedNodeId { get; set; }
         public int[] Checks { get; set; }
         public bool StartAudit { get; set; }
