+++ src/SeoToolkit.Umbraco.SiteAudit/assets/src/workspaces/SiteAuditDetailContext.ts
@@ -60,6 +60,9 @@ export default class SiteAuditDetailContext
   loadData(unique: string) {
     this.#repository.get(Number.parseInt(unique)).then((res) => {
       this.#model.update(res.data!);
+      if (res.data?.status === 'Finished'){
+        this.runHeartbeat = false;
+      }
     });
   }
